refactor the new views into files
