+++ Improvify/view/DateAndHabitsView.swift
@@ -1,0 +1,45 @@
+//
+//  DateAndHabitsView.swift
+//  Improvify
+//
+//  Created by Maximiliano Paris Gaete on 1/16/25.
+//
+import SwiftUI
+
+struct DateAndHabitsView: View {
+    var habitsManager: HabitsManager
+    var date: Date
+    
+    var body: some View {
+        List {
+            Section {
+                HStack {
+                    Image(systemName: "arrow.left")
+                        .onTapGesture {
+                            habitsManager.moveDayBackward()
+                        }
+                    
+                    Spacer()
+                    Text("\(habitsManager.dateFormatter.string(from: date))")
+                        .onTapGesture {
+                            habitsManager.goToToday()
+                        }
+                    Spacer()
+                    
+                    Image(systemName: "arrow.right")
+                        .onTapGesture {
+                            habitsManager.moveDayForward()
+                        }
+                }
+            }
+            
+            Section {
+                ForEach(habitsManager.habits, id: \.self) { habit in
+                    HabitRow(habitsManager: habitsManager, habit: habit, date: date)
+                    
+                }
+            }
+        }
+        .id(date)
+    }
+}
